requirements.txt
boto3==1.9.237
botocore==1.12.237
Click==7.0
docutils==0.15.2
Flask==1.1.1
flask-lambda-support==0.1.5
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
